move FilesToProcess to pipeline
